Update map.html
Q4, condition the quesiton based on cubf==1000 or not
Since the field cubf is non empty, so:
if cubf ==1000, display questions for MURBs
if not, then display questions for community center/schools/other

Q5, condition the question based on construction year.
if constructed before 2000, display subquestions designed for building before 2000
else if constructed after 2000, display nothing in section3
else if no field (construction year not existed) in .csv, then ask user to choose whether it is constructed before or after 2000.

Q6, 7, display only when the field is missing in csv.

Q12, 14, 16, under discussion. 

code is not cleaned in case of future modifications.